Fix work authorization not appearing due to missing default values
The work authorization field wasn't showing because templates had incomplete default values for the personal object. Fixed by:

1. Added nationality to CV store's createDefaultCV function
2. Updated all template default values to include nationality: ''
   - Harvard Template
   - Dublin Tech Template
   - Dublin Pharma Template
   - Irish Graduate Template
   - Irish Finance Template
   - Classic Template

3. Added debug logging to track personal data

The issue was that when destructuring personal data, the default object didn't include nationality, so it would be undefined even when data was present.

Now work authorization will properly display in all templates when entered in the form.